Refactor: rename AirtableResourceNotFound to LearningContentResourceNotFound
