--- api/lib/infrastructure/datasources/learning-content/AirtableResourceNotFound.js
@@ -1,7 +1,0 @@
-class AirtableResourceNotFound extends Error {
-  constructor() {
-    super();
-  }
-}
-
-module.exports = AirtableResourceNotFound;
