+++ api/lib/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js
@@ -1,0 +1,7 @@
+class LearningContentResourceNotFound extends Error {
+  constructor() {
+    super();
+  }
+}
+
+module.exports = LearningContentResourceNotFound;
